Try to detect duplicate paths while adding new files to database
This doesn't necessarily work, even though the SQL is right. The path of
the file is relative to which directory user chooses to import. For the
same file it can be the immediately enclosing directory or few levels up
the hierarchy - i.e. for the same file path won't necessarily be unique
(it can be subpath or superpath)