Avoid reading 1000s of null frames with a safety limit on iteration
For some reason - probably my lack of understanding of spec - there seem
to appear 1000s of empty frames. Most of the relevant frames are always
at the beginning of the tag, therefore we can save time by avoiding
these unnecessary iterations (until we understand the reason behind them
and fix them correctly). This patch fixes that.