Changing FileSystem type to TEMPORARY instead of PERSISTENT
Has the big advantage that it does not require user to run the browser
with any special flags, also doesn't require user's permission for quota
allocation. On the downside, the files can be deleted by the browser at
will. Since our usage of file system is caching purposes, we can live
with this limitation.