Fixes the problem of 3DTin logo not being clickable
The reason was the element carrying "Artwork powered by" was spanning
entire width of the page and hence was overlapping on the 3DTin icon.
Therefore hovering over the 3DTin icon wasn't turning the cursor to
pointer or opening the link when clicked. It was only effective only
certain portion of the icon.

This fix limits the width of google_branding's child elements within
200px, so that they don't interfere with rest of the elements in the
footer.